@@ -3,17 +3,19 @@ package com.wzc.constraintlayoutexamples
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
-import kotlinx.android.synthetic.main.activity_barrier.*
+import com.wzc.constraintlayoutexamples.databinding.ActivityBarrierBinding
 
 class BarrierActivity : AppCompatActivity() {
     private var changed: Boolean = false
+    private lateinit var binding: ActivityBarrierBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_barrier)
-        button_change.setOnClickListener {
+        binding = ActivityBarrierBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        binding.buttonChange.setOnClickListener {
 
-            val buttonALp = button_a.layoutParams as ConstraintLayout.LayoutParams
-            val buttonBLp = button_b.layoutParams as ConstraintLayout.LayoutParams
+            val buttonALp = binding.buttonA.layoutParams as ConstraintLayout.LayoutParams
+            val buttonBLp = binding.buttonB.layoutParams as ConstraintLayout.LayoutParams
             if (changed) {
                 buttonALp.width = 100.dp2px().toInt()
                 buttonBLp.width = 50.dp2px().toInt()
@@ -22,8 +24,8 @@ class BarrierActivity : AppCompatActivity() {
                 buttonBLp.width = 150.dp2px().toInt()
             }
             changed = !changed
-            button_a.layoutParams = buttonALp
-            button_b.layoutParams = buttonBLp
+            binding.buttonA.layoutParams = buttonALp
+            binding.buttonB.layoutParams = buttonBLp
         }
     }
 }
